+++ tests/AdminActor.test.ts
@@ -1,10 +1,6 @@
-import { dbInsertAdmin, dbSelectAdmin } from "@/lib/data/dbAdmins";
 import { withDb } from "./_db_helpers";
-import { AdminPii, encryptAdminPii } from "@/lib/admin/admin-pii";
-import { Admin, AdminSpec } from "@/lib/data/models";
-import { HarnessEncryptionService, HarnessHashService } from "./_harness";
-import { AdminActor, AdminActorConfig } from "@/lib/admin/admin-actor";
-import { Pool } from "pg";
+import { AdminActor } from "@/lib/admin/admin-actor";
+import { createAdmin, getAdminActorConfig, adminPii } from "./_fixtures";
 
 describe("AdminActor", () => {
   it("can retrieve its own actor data from the database", async () => {
@@ -26,39 +22,3 @@ describe("AdminActor", () => {
     });
   });
 });
-
-const emailHashService = new HarnessHashService();
-const piiEncryptionService = new HarnessEncryptionService();
-
-function getAdminActorConfig(db: Pool): AdminActorConfig {
-  return {
-    dbPool: db,
-    emailHashService: emailHashService,
-    piiEncryptionService: piiEncryptionService,
-  };
-}
-
-async function createAdmin(idx: number, db: Pool): Promise<Admin> {
-  const spec = await adminSpec(1);
-  const gen = await dbInsertAdmin(db, spec);
-  const admin = await dbSelectAdmin(db, gen.adminId);
-  if (admin === null) {
-    fail("Failed to retrieve admin");
-  }
-  return admin;
-}
-
-async function adminSpec(idx: number): Promise<AdminSpec> {
-  const pii = adminPii(idx);
-  const adminEncryptedPii = await encryptAdminPii(pii, piiEncryptionService);
-  const adminHashedEmail = await emailHashService.getHashHex(pii.adminEmail);
-  return { adminHashedEmail, adminEncryptedPii };
-}
-
-function adminPii(idx: number): AdminPii {
-  return {
-    adminEmail: `admin${idx}@admin.com`,
-    adminName: `Admin ${idx}`,
-    adminPhoneNumber: `+65 ${10000000 + idx}`,
-  };
-}
